Controller and DAL for Product created. Product Model adjusted to match db setup

+++ OrientationApi/OrientationApi/Models/Product.cs
@@ -9,6 +9,9 @@ namespace OrientationApi.Models
     {
         public int ProductId { get; set; }
         public string ProductName { get; set; }
-        public double ProductPrice { get; set; }
+        public int ProductPrice { get; set; }
+        //Price is an int just because of the way the database was set up
+        //We can change this to a double later...
+        //...we will make sure the methods work, first.
     }
 }
